Update Remove-VpnConnections.ps1
We are running into the known duplicate clients issue with device tunnel and thought this script would be cleaner than restarting the service regularly. In the original version of the script, since the Disconnect-VpnUser command targets the username, it closes out all connections with that name, not just ones that we consider stale. The solution I recommend is disconnecting by HostIpAddress instead of Username.

In addition, I added some code to the end to remove duplicate entries based on username (device name in this case due to device tunneling) and leaving only the newest ones based on connection duration.